+++ tests/unit/processor/grokker/test_grokker_rule.py
@@ -212,6 +212,30 @@ class TestGrokkerRule:
                 ValueError,
                 "must match regex",
             ),
+            (
+                {
+                    "filter": "message",
+                    "grokker": {
+                        "mapping": {
+                            "message": "(?<group>(TOO(%{SPACE}(M(AN)Y)%{SPACE})PARENTHESES))"
+                        }
+                    },
+                },
+                InvalidRuleDefinitionError,
+                r"The resolved grok pattern.*is not valid",
+            ),
+            (
+                {
+                    "filter": "message",
+                    "grokker": {
+                        "mapping": {
+                            "message": "(?<group>(NOT-TOO(%{SPACE}(MANY)%{SPACE})PARENTHESES))"
+                        }
+                    },
+                },
+                None,
+                None,
+            ),
         ],
     )
     def test_create_from_dict_validates_config(self, rule, error, message):