+++ CHANGELOG.md
@@ -11,6 +11,12 @@ representation is not completely lucene compatible due to non-existing regex fun
 ### Bugfix
 
 * Fix error handling of FieldManager if no mapped source field exists in the event.
+* Fix Grokker such that only the first grok pattern match is applied instead of all matching
+pattern
+* Fix Grokker such that nested parentheses in oniguruma pattern are working (3 levels are supported
+now)
+* Fix Grokker such that two or more oniguruma can point to the same target. This ensures
+grok-pattern compatibility with the normalizer and other grok tools
 
 ## v6.3.0
 ### Features